+++ src/main/java/com/jifenke/lepluslive/merchant/domain/entities/MerchantProtocol.java
@@ -1,6 +1,5 @@
 package com.jifenke.lepluslive.merchant.domain.entities;
 
-import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -20,36 +19,10 @@ public class MerchantProtocol {
   private Long id;
 
 
-  private String picture;                     // 图片 URL
+  private String picture;
 
   @ManyToOne
-  private Merchant merchant;                  // 商户id
-
-
-  // 图片类型: 1-平台受理服务协议书  2-商户基础资料表 3-中汇支付收单特约商户信息调查表  4-结算授权书
-  @Column(name = "protocol_type", length = 20, columnDefinition = "BIGINT")
-  private Long protocolType;
-
-  @Column(name = "pic_index", length = 20, columnDefinition = "BIGINT")
-  private Long picIndex;                         // 位置索引: 从0开始
-  @Column(name = "pic_state", length = 20, columnDefinition = "BIGINT")
-  private Long picState;                         // 图片状态: 0-正常 1-失效
-
-  public Long getPicIndex() {
-    return picIndex;
-  }
-
-  public void setPicIndex(Long picIndex) {
-    this.picIndex = picIndex;
-  }
-
-  public Long getPicState() {
-    return picState;
-  }
-
-  public void setPicState(Long picState) {
-    this.picState = picState;
-  }
+  private Merchant merchant;
 
   public Long getId() {
     return id;
@@ -73,13 +46,5 @@ public class MerchantProtocol {
 
   public void setMerchant(Merchant merchant) {
     this.merchant = merchant;
-  }
-
-  public Long getProtocolType() {
-    return protocolType;
-  }
-
-  public void setProtocolType(Long protocolType) {
-    this.protocolType = protocolType;
   }
 }
